Explicit cast to uint256

+++ contracts/CTokenInterfaces.sol
@@ -29,7 +29,7 @@ contract CTokenStorage {
 
     // Maximum borrow rate that can ever be applied (.0005% / block)
     /// @dev modified to use a per second borrow rate for zksync compatibility
-    uint internal constant borrowRateMaxMantissa = 0.0005e16 / 12;
+    uint internal constant borrowRateMaxMantissa = 0.0005e16 / uint256(12);
 
     // Maximum fraction of interest that can be set aside for reserves
     uint internal constant reserveFactorMaxMantissa = 1e18;
